CM-1631: Moved REST API data model classes into implementation details. Exposed ExperimentMetadata public DTO object.
