+++ comet-java-client/src/main/java/ml/comet/experiment/Experiment.java
@@ -2,11 +2,12 @@ package ml.comet.experiment;
 
 import ml.comet.experiment.context.ExperimentContext;
 import ml.comet.experiment.impl.asset.AssetType;
-import ml.comet.experiment.model.ExperimentAssetLink;
-import ml.comet.experiment.model.ExperimentMetadataRest;
-import ml.comet.experiment.model.GitMetadata;
-import ml.comet.experiment.model.GitMetadataRest;
-import ml.comet.experiment.model.ValueMinMaxDto;
+import ml.comet.experiment.impl.model.ExperimentAssetLink;
+import ml.comet.experiment.impl.model.ExperimentMetadataRest;
+import ml.comet.experiment.impl.model.GitMetadata;
+import ml.comet.experiment.impl.model.GitMetadataRest;
+import ml.comet.experiment.impl.model.ValueMinMaxDto;
+import ml.comet.experiment.model.ExperimentMetadata;
 
 import java.io.File;
 import java.util.List;
@@ -65,7 +66,7 @@ public interface Experiment extends AutoCloseable {
      *
      * @return the metadata associated with experiment.
      */
-    ExperimentMetadataRest getMetadata();
+    ExperimentMetadata getMetadata();
 
     /**
      * Get experiment git metadata.
