+++ comet-java-client/src/main/java/ml/comet/experiment/impl/BaseExperimentAsync.java
@@ -11,13 +11,13 @@ import ml.comet.experiment.context.ExperimentContext;
 import ml.comet.experiment.impl.asset.Asset;
 import ml.comet.experiment.impl.asset.RemoteAsset;
 import ml.comet.experiment.impl.utils.AssetUtils;
-import ml.comet.experiment.model.GitMetadata;
-import ml.comet.experiment.model.HtmlRest;
-import ml.comet.experiment.model.LogDataResponse;
-import ml.comet.experiment.model.LogOtherRest;
-import ml.comet.experiment.model.MetricRest;
-import ml.comet.experiment.model.OutputUpdate;
-import ml.comet.experiment.model.ParameterRest;
+import ml.comet.experiment.impl.model.GitMetadata;
+import ml.comet.experiment.impl.model.HtmlRest;
+import ml.comet.experiment.impl.model.LogDataResponse;
+import ml.comet.experiment.impl.model.LogOtherRest;
+import ml.comet.experiment.impl.model.MetricRest;
+import ml.comet.experiment.impl.model.OutputUpdate;
+import ml.comet.experiment.impl.model.ParameterRest;
 import org.slf4j.Logger;
 
 import java.io.File;
