+++ comet-java-client/src/main/java/ml/comet/experiment/impl/CometApiImpl.java
@@ -9,8 +9,8 @@ import ml.comet.experiment.impl.config.CometConfig;
 import ml.comet.experiment.impl.http.Connection;
 import ml.comet.experiment.impl.http.ConnectionInitializer;
 import ml.comet.experiment.impl.utils.CometUtils;
-import ml.comet.experiment.model.ExperimentMetadataRest;
-import ml.comet.experiment.model.RestProject;
+import ml.comet.experiment.impl.model.ExperimentMetadataRest;
+import ml.comet.experiment.impl.model.RestProject;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
