+++ comet-java-client/src/main/java/ml/comet/experiment/impl/RestApiClient.java
@@ -12,30 +12,30 @@ import ml.comet.experiment.impl.constants.QueryParamName;
 import ml.comet.experiment.impl.http.Connection;
 import ml.comet.experiment.impl.utils.AssetUtils;
 import ml.comet.experiment.impl.utils.JsonUtils;
-import ml.comet.experiment.model.AddExperimentTagsRest;
-import ml.comet.experiment.model.AddGraphRest;
-import ml.comet.experiment.model.CreateExperimentRequest;
-import ml.comet.experiment.model.CreateExperimentResponse;
-import ml.comet.experiment.model.ExperimentAssetListResponse;
-import ml.comet.experiment.model.ExperimentMetadataRest;
-import ml.comet.experiment.model.ExperimentStatusResponse;
-import ml.comet.experiment.model.ExperimentTimeRequest;
-import ml.comet.experiment.model.GetExperimentsResponse;
-import ml.comet.experiment.model.GetGraphResponse;
-import ml.comet.experiment.model.GetHtmlResponse;
-import ml.comet.experiment.model.GetOutputResponse;
-import ml.comet.experiment.model.GetProjectsResponse;
-import ml.comet.experiment.model.GetWorkspacesResponse;
-import ml.comet.experiment.model.GitMetadata;
-import ml.comet.experiment.model.GitMetadataRest;
-import ml.comet.experiment.model.HtmlRest;
-import ml.comet.experiment.model.LogDataResponse;
-import ml.comet.experiment.model.LogOtherRest;
-import ml.comet.experiment.model.MetricRest;
-import ml.comet.experiment.model.MinMaxResponse;
-import ml.comet.experiment.model.OutputUpdate;
-import ml.comet.experiment.model.ParameterRest;
-import ml.comet.experiment.model.TagsResponse;
+import ml.comet.experiment.impl.model.AddExperimentTagsRest;
+import ml.comet.experiment.impl.model.AddGraphRest;
+import ml.comet.experiment.impl.model.CreateExperimentRequest;
+import ml.comet.experiment.impl.model.CreateExperimentResponse;
+import ml.comet.experiment.impl.model.ExperimentAssetListResponse;
+import ml.comet.experiment.impl.model.ExperimentMetadataRest;
+import ml.comet.experiment.impl.model.ExperimentStatusResponse;
+import ml.comet.experiment.impl.model.ExperimentTimeRequest;
+import ml.comet.experiment.impl.model.GetExperimentsResponse;
+import ml.comet.experiment.impl.model.GetGraphResponse;
+import ml.comet.experiment.impl.model.GetHtmlResponse;
+import ml.comet.experiment.impl.model.GetOutputResponse;
+import ml.comet.experiment.impl.model.GetProjectsResponse;
+import ml.comet.experiment.impl.model.GetWorkspacesResponse;
+import ml.comet.experiment.impl.model.GitMetadata;
+import ml.comet.experiment.impl.model.GitMetadataRest;
+import ml.comet.experiment.impl.model.HtmlRest;
+import ml.comet.experiment.impl.model.LogDataResponse;
+import ml.comet.experiment.impl.model.LogOtherRest;
+import ml.comet.experiment.impl.model.MetricRest;
+import ml.comet.experiment.impl.model.MinMaxResponse;
+import ml.comet.experiment.impl.model.OutputUpdate;
+import ml.comet.experiment.impl.model.ParameterRest;
+import ml.comet.experiment.impl.model.TagsResponse;
 
 import java.io.File;
 import java.util.Collections;
