+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/CreateExperimentResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,13 +7,14 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class OutputLine {
-    private String output;
-    private boolean stderr = false;
-    private long localTimestamp;
-    private Long offset;
+public class CreateExperimentResponse {
+    private String experimentKey;
+    private String workspaceName;
+    private String projectName;
+    private String link;
+    private String name;
 }
