+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/ExperimentAssetListResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,16 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GitMetadata {
-    private String experimentKey;
-    private String user;
-    private String root;
-    private String branch;
-    private String parent;
-    private String origin;
+public class ExperimentAssetListResponse {
+    private List<ExperimentAssetLink> assets;
 }
