+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/ExperimentStatusResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,13 +6,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
+/**
+ * Describes response received when experiment set status sent.
+ */
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetExperimentsResponse {
-    List<ExperimentMetadataRest> experiments;
+public class ExperimentStatusResponse {
+    private long isAliveBeatDurationMillis;
+    private long gpuMonitorIntervalMillis;
+    private long cpuMonitorIntervalMillis;
 }
