+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/ExperimentTimeRequest.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,16 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/**
- * Describes response received when experiment set status sent.
- */
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ExperimentStatusResponse {
-    private long isAliveBeatDurationMillis;
-    private long gpuMonitorIntervalMillis;
-    private long cpuMonitorIntervalMillis;
+public class ExperimentTimeRequest {
+    private String experimentKey;
+    private Long startTimeMillis;
+    private Long endTimeMillis;
 }
