+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/GetExperimentsResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,13 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ExperimentTimeRequest {
-    private String experimentKey;
-    private Long startTimeMillis;
-    private Long endTimeMillis;
+public class GetExperimentsResponse {
+    List<ExperimentMetadataRest> experiments;
 }
