+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/HtmlRest.java
@@ -1,7 +1,8 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -11,9 +12,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class LogOtherRest {
-    private String key;
-    private String value;
+public class HtmlRest {
+    @JsonProperty(required = true)
+    private String html;
     private String experimentKey;
+    private Boolean override = false;
     private Long timestamp;
 }
