+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/MinMaxResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,15 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class CreateExperimentResponse {
-    private String experimentKey;
-    private String workspaceName;
-    private String projectName;
-    private String link;
-    private String name;
+public class MinMaxResponse {
+    List<ValueMinMaxDto> values;
 }
