+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/ParameterRest.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,14 +6,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AddExperimentTagsRest {
+public class ParameterRest {
+    private String parameterName;
+    private String parameterValue;
+    private Long step;
     private String experimentKey;
-    private List<String> addedTags;
+    private Long timestamp;
+    private String context;
 }
