+++ comet-java-client/src/main/java/ml/comet/experiment/impl/model/TagsResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.model;
+package ml.comet.experiment.impl.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,16 +6,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ParameterRest {
-    private String parameterName;
-    private String parameterValue;
-    private Long step;
-    private String experimentKey;
-    private Long timestamp;
-    private String context;
+public class TagsResponse {
+    private List<String> tags;
 }
