--- comet-java-client/src/main/java/ml/comet/experiment/model/HtmlRest.java
@@ -1,21 +1,0 @@
-package ml.comet.experiment.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class HtmlRest {
-    @JsonProperty(required = true)
-    private String html;
-    private String experimentKey;
-    private Boolean override = false;
-    private Long timestamp;
-}
