+++ comet-java-client/src/main/java/ml/comet/experiment/model/package-info.java
@@ -1,4 +1,5 @@
 /**
- * Contains all classes used to model Comet REST API data transfer objects.
+ * Contains all classes to encapsulate data model objects used by the Comet experiment which exposed through the
+ * public API.
  */
 package ml.comet.experiment.model;
